Fix L button activation behavior
WM_MOUSEACTIVATE procedure always ate mouse L button messages
when activated by the button.
However, even if you clicked the close button etc.
when the window was not active, it did not respond.

This commit makes the procedure eats the messages
only in the client area.
Therefore, left clicking on the non-client area
when the window is inactive is enabled.