Fix class pointer bug
Base class's `this` pointer is *not* derived class's pointer.
In the case of simple class inheritance,
even if reinterpret_cast from base class pointer
to derived class pointer, it may appear to be working.
However, it is a dangerous mistake.

Unfortunately, since the last argument of `CreateWindow ()` is
`void *` type, the original type (base class's pointer) is lost.
When receiving the pointer in the window procedure,
there is no type so we need to cast it.
However, casting to the derived class's pointer
with `reinterpret_cast` raises a problem
because it is originally the base class's pointer.
If simple classes, it may seem like it works.
However, for a complicated classes,
it may raise segmentation fault.

This commit makes `window_class<Derived>::create_window ()`
casts `this` pointer to derived class's pointer
with `dynamic_cast` before passing it to `CreateWindow ()`.
It is casted to `void *` and passed to the window procedure,
but since correctly casts to the derived class pointer
with `reinterpret_cast`, it works correctly.